sec status colours everywhere
