+++ src/lib/components/CharacterList.svelte
@@ -1,38 +1,8 @@
 <script>
 	import { Avatar } from 'flowbite-svelte';
+	import { secStatusColor } from '$lib/utils/secStatus';
 
 	let { pilots } = $props();
-
-	const clamp = (v, min, max) => Math.min(max, Math.max(min, v));
-	const lerp = (a, b, t) => a + (b - a) * t;
-	const hslToString = (h, s, l) => `hsl(${Math.round(h)} ${Math.round(s)}% ${Math.round(l)}%)`;
-	const mixHsl = (a, b, t) => [lerp(a[0], b[0], t), lerp(a[1], b[1], t), lerp(a[2], b[2], t)];
-
-	function secStatusColor(value) {
-		const v = clamp(value, -10, 5);
-		// endpoints: red (-10), gray (0), green (5)
-		const red = [0, 75, 45];
-		const gray = [0, 0, 50];
-		const green = [140, 65, 40];
-
-		if (v === 0) {
-			return hslToString(...gray);
-		}
-
-		if (v < 0) {
-			// Ease to be less gray near small negatives
-			const tLinear = (v + 10) / 10; // -10 -> 0, 0 -> 1
-			const t = Math.pow(tLinear, 3); // push towards red for values near 0
-			const [h, s, l] = mixHsl(red, gray, t);
-			return hslToString(h, s, l);
-		} else {
-			// Ease to be less gray near small positives
-			const tLinear = v / 5; // 0 -> 0, 5 -> 1
-			const t = Math.pow(tLinear, 1 / 3); // pull towards green for values near 0
-			const [h, s, l] = mixHsl(gray, green, t);
-			return hslToString(h, s, l);
-		}
-	}
 </script>
 
 <h1 class="ms-2 text-xl font-bold">Pilots</h1>
